shift all basic functions out of __main__.py
